list-selector-grid-item: Allow control-click for selection mode.
Currently entering selection mode is possible through clicking the 'selection
mode' button or by right-clicking a grid item.

According to the style guide it's recomended to enable selection mode by
control/shift-clicking an item.

To remedy this we now allow ctrl-left-click on grid items to enter
selection mode and mark them as selected.

https://bugzilla.gnome.org/show_bug.cgi?id=778675