manager: remove old signals on provider-removal
Currently, the signals connected to providers are not disconnected
after the provider is removed.It doesn't cause any problem with
providers which are built-in but with providers that work as a plugin
e.g Todo.txt this would cause weird behaviour.
Removing all handlers on the signals that were connected initially,
solves the problem.

https://bugzilla.gnome.org/show_bug.cgi?id=781079