task-row: improve aestethics of task rows
The current behavior of task rows isn't really good. Because
it is transparent, the contents of the task row look odd with
crazy background colors. And the task priority colors become
basically useless.

This commit introduces a new layout to these task rows, where
priorities are shown in the beginning of the row and the background
color is completely opaque.